Changed TestDataSource to ITestDataSource.
