+++ src/TestFramework/MSTest.Core/MSTest.Core.csproj
@@ -36,7 +36,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="Attributes\DynamicDataAttribute.cs" />
-    <Compile Include="Attributes\TestDataSourceAttribute.cs" />
+    <Compile Include="Interfaces\ITestDataSource.cs" />
     <Compile Include="Interfaces\ITestMethod.cs" />
     <Compile Include="Internal\Helper.cs" />
     <Compile Include="DataAccessMethod.cs" />
