+++ test/UnitTests/MSTest.Core.Unit.Tests/Attributes/DynamicDataAttributeTests.cs
@@ -32,6 +32,7 @@ namespace Microsoft.VisualStudio.TestPlatform.TestFramework.UnitTests.Attributes
         {
             this.dummyTestClass = new DummyTestClass();
             this.testMethodInfo = this.dummyTestClass.GetType().GetTypeInfo().GetDeclaredMethod("TestMethod1");
+            this.dynamicDataAttribute = new DynamicDataAttribute("ReusableTestDataProperty");
         }
 
         [TestFrameworkV1.TestMethod]
@@ -110,6 +111,39 @@ namespace Microsoft.VisualStudio.TestPlatform.TestFramework.UnitTests.Attributes
             };
 
             ActionUtility.ActionShouldThrowExceptionOfType(action, typeof(ArgumentNullException));
+        }
+
+        [TestFrameworkV1.TestMethod]
+        public void GetDisplayNameShouldReturnDisplayName()
+        {
+            var data = new object[] { 1, 2, 3 };
+
+            var displayName = this.dynamicDataAttribute.GetDisplayName(this.testMethodInfo, data);
+            Assert.AreEqual("TestMethod1 (1,2,3)", displayName);
+        }
+
+        [TestFrameworkV1.TestMethod]
+        public void GetDisplayNameShouldReturnEmptyStringIfDataIsNull()
+        {
+            var displayName = this.dynamicDataAttribute.GetDisplayName(this.testMethodInfo, null);
+            Assert.IsNull(displayName);
+        }
+
+        [TestFrameworkV1.TestMethod]
+        public void GetDisplayNameShouldThrowIfDataHasNullValues()
+        {
+            var data = new string[] { "value1", "value2", null };
+            var data1 = new string[] { null, "value1", "value2" };
+            var data2 = new string[] { "value1", null, "value2" };
+
+            var displayName = this.dynamicDataAttribute.GetDisplayName(this.testMethodInfo, data);
+            Assert.AreEqual("TestMethod1 (value1,value2,)", displayName);
+
+            displayName = this.dynamicDataAttribute.GetDisplayName(this.testMethodInfo, data1);
+            Assert.AreEqual("TestMethod1 ()", displayName);
+
+            displayName = this.dynamicDataAttribute.GetDisplayName(this.testMethodInfo, data2);
+            Assert.AreEqual("TestMethod1 (value1,,value2)", displayName);
         }
     }
