+++ test/UnitTests/MSTest.Core.Unit.Tests/MSTest.Core.Unit.Tests.csproj
@@ -71,7 +71,6 @@
     <Compile Include="Attributes\ExpectedExceptionAttributeTests.cs" />
     <Compile Include="Attributes\ExpectedExceptionBaseAttributeTests.cs" />
     <Compile Include="Attributes\DynamicDataAttributeTests.cs" />
-    <Compile Include="Attributes\TestDataSourceAttributeTests.cs" />
     <Compile Include="GenericParameterHelperTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
